Adds "Join as Antag" to Lobby Choices
Adds a join as antag button for late join antags. Antags are to be limited to current pop number and police numbers to keep balance.

Hopefully this means we can start having a bit of pizazz but more planning on how to keep this decent is required. Will expand on what antag types and what they have access to later. Putting this here for now.